requirements.txt
pettingzoo[classic]>=1.24.0
pygame>=2.5.0
numpy>=1.24.0
gymnasium>=0.29.0
pyyaml>=6.0
matplotlib>=3.7.0
streamlit>=1.28.0
setuptools>=65.5.0
wheel>=0.40.0
# Week 3 Dependencies
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
osmnx>=1.7.0
networkx>=3.1
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
Pillow>=10.0.0